+++ src/app/core/models/event-type.enum.ts
@@ -1,0 +1,15 @@
+export enum EventType {
+  SIMPLE = 'SIMPLE',
+  SPLITTED = 'SPLITTED',
+  SUPEREVENT = 'SUPEREVENT',
+}
+
+const EVENT_TYPES: EventType[] = [
+  EventType.SIMPLE,
+  EventType.SPLITTED,
+  EventType.SUPEREVENT,
+];
+
+const EVENT_TYPE_DEFAULT = EventType.SIMPLE;
+
+export { EVENT_TYPES, EVENT_TYPE_DEFAULT };