+++ src/app/core/pipes/eventType.pipe.ts
@@ -1,0 +1,28 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+import { EventType } from '@models/event-type.enum';
+
+@Pipe({
+  name: 'eventType'
+})
+export class EventTypePipe implements PipeTransform {
+
+  transform(value: string): string {
+    switch(value) {
+
+      case EventType.SIMPLE:
+        return 'EVENTO SENCILLO';
+
+      case EventType.SPLITTED:
+        return 'EVENTO DESGLOSADO';
+
+      case EventType.SUPEREVENT:
+        return 'SUPER EVENTO';
+
+
+      default:
+        return '*** ERROR ***';
+    }
+  }
+
+}