+++ src/app/features/events/event-view/event-view.component.ts
@@ -7,6 +7,7 @@ import { environment } from '@environments/environment';
 import { AuthService } from '@auth/auth.service';
 import { Base, IBase } from '@models/base';
 import { IEvent, Event } from '@models/event';
+import { EventType } from '@models/event-type.enum';
 import { IUser } from '@models/user';
 import { IAppointment, Appointment } from '@models/appointment';
 import { AppointmentType } from '@models/appointment-type';
@@ -39,8 +40,9 @@ export class EventViewComponent implements OnInit, OnDestroy {
   public subEvent: IEvent;
   public appointmentSubEvent: IAppointment;
 
-
   readonly SECTION_BLANK: Base = Base.InitDefault();
+  readonly EVENT_TYPE_SPLITTED = EventType.SPLITTED;
+  readonly EVENT_TYPE_SUPEREVENT = EventType.SUPEREVENT;
   private listOfObservers: Array<Subscription> = [];
 
   constructor(